Reverted unnecessary changes
