+++ libs/core/src/framework/types.ts
@@ -7,7 +7,6 @@ import { Events, events } from '../schemas';
 export const INVALID_INPUT_ERROR = 'Invalid Input Error';
 export const INVALID_ACTOR_ERROR = 'Invalid Actor Error';
 export const INVALID_STATE_ERROR = 'Invalid State Error';
-export const ENDPOINT_DISABLED = 'Endpoint Off Error';
 
 /**
  * Deep partial utility
@@ -76,20 +75,6 @@ export class InvalidState extends Error {
   ) {
     super(message);
     this.name = INVALID_STATE_ERROR;
-  }
-}
-
-/**
- * Endpoint off error - The api is disabled via feature flags
- */
-export class EndpointOff extends Error {
-  constructor(
-    message: string,
-    public readonly state?: unknown,
-    public readonly payload?: unknown,
-  ) {
-    super(message);
-    this.name = ENDPOINT_DISABLED;
   }
 }
