+++ libs/model/src/middleware/guards.ts
@@ -1,5 +1,4 @@
-import { EndpointOff, InvalidState } from '@hicommonwealth/core';
-import { OpenFeature } from '@openfeature/server-sdk';
+import { InvalidState } from '@hicommonwealth/core';
 
 export const mustExist = <T>(subject: string, state?: T | null): state is T => {
   if (!state) throw new InvalidState(`${subject} must exist`, state);
@@ -8,15 +7,4 @@ export const mustExist = <T>(subject: string, state?: T | null): state is T => {
 
 export const mustNotExist = <T>(subject: string, state?: T | null) => {
   if (state) throw new InvalidState(`${subject} must not exist`, state);
-};
-
-export const behindFeatureFlag = async (flag: string) => {
-  if (process.env.NODE_ENV === 'test') {
-    return;
-  }
-
-  const client = OpenFeature.getClient();
-  if (!(await client.getBooleanValue(flag, false))) {
-    throw new EndpointOff('This endpoint has been disabled', flag);
-  }
 };
